Fixed a warning.

+++ Problems/Two_Sum/Two_Sum.cpp
@@ -12,7 +12,7 @@ public:
 		std::vector<int> result;
 		size_t nums_size = nums.size();
 
-		for (int i = 0; i < nums_size; i++)
+		for (size_t i = 0; i < nums_size; i++)
 		{
 			if (target != 0)
 			{
@@ -22,7 +22,7 @@ public:
 				}
 			}
 
-			for (int j = i + 1; j < nums_size; j++)
+			for (size_t j = i + 1; j < nums_size; j++)
 			{
 				if (nums[i] + nums[j] == target)
 				{
